Backend Bedrock Service をリファクタリングし、テスト容易性を高めておく
