+++ src/main/api/index.ts
@@ -1,8 +1,8 @@
 import express, { Request, Response } from 'express'
 import cors from 'cors'
-import { bedrock, CallConverseAPIProps } from './bedrock'
 import { RequestHandler, NextFunction } from 'express'
 import { RetrieveAndGenerateCommandInput } from '@aws-sdk/client-bedrock-agent-runtime'
+import { bedrock, CallConverseAPIProps } from './bedrock'
 
 interface PromiseRequestHandler {
   (req: Request, res: Response, next: NextFunction): Promise<unknown>
